Shrink the kernel  a little bit more.

+++ src/opencl/sha512_kernel.cl
@@ -22,7 +22,7 @@ inline void _memcpy(               uint32_t * dest,
         *dest++ = *src++;
 }
 
-inline uint64_t sha512_block(sha512_ctx * ctx) {
+inline uint32_t sha512_block(sha512_ctx * ctx) {
     uint64_t a = H0;
     uint64_t b = H1;
     uint64_t c = H2;
@@ -89,7 +89,7 @@ void kernel_crypt_raw(__global   const uint32_t  * keys_buffer,
     CLEAR_BUFFER_64_FAST(ctx.buffer->mem_64, ctx.buflen + 1);
 
     /* Run the collected hash value through SHA512. Return parcial results */
-    out_buffer[gid] = (uint32_t) sha512_block(&ctx);
+    out_buffer[gid] = sha512_block(&ctx);
 }
 
 __kernel
@@ -140,7 +140,7 @@ void kernel_crypt_xsha(__constant sha512_salt     * salt,
     CLEAR_BUFFER_64_FAST(ctx.buffer->mem_64, ctx.buflen + 1);
 
     /* Run the collected hash value through SHA512. Return parcial results */
-    out_buffer[gid] = (uint32_t) sha512_block(&ctx);
+    out_buffer[gid] = sha512_block(&ctx);
 }
 
 __kernel
